requirements.txt
Django==5.2.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
numpy==2.3.2
pandas==2.3.1
PyYAML==6.0.2
scikit_learn==1.7.1
mysqlclient==2.2.4 